api/requirements.txt

flask==2.0.1
flask-cors==3.0.10
openai==1.3.0
python-dotenv==0.19.0
requests==2.26.0
